Fixed Object Falling Through Floors, Trigger Syncing
Disabled RBNetComp, along with a few additional checks which seems to have fixed the issue of the object falling through the floor when being picked up.
Added logic for the TriggerNetComp to detect what type of trigger its attached to. Allowing all types of triggers to be synced across the network.
Attempted to fix Launchpad.

Bugs:-
Player holding the object will teleport as the client struggles to update the positions of both the player and the object